move new test into a separate file
